Update `timeout_s` type hint from int to second

+++ python/ray/train/v2/_internal/execution/checkpoint/checkpoint_manager.py
@@ -520,7 +520,7 @@ class CheckpointManager(_CheckpointManager, ReportCallback, WorkerGroupCallback)
         self,
         current_report_index: int,
         consistency_mode: CheckpointConsistencyMode = CheckpointConsistencyMode.VALIDATED,
-        timeout_s: int = -1,
+        timeout_s: float = -1,
     ) -> List[ReportedCheckpoint]:
         """Get all the reported checkpoints so far.
